+++ helpers/helpers.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace rr;
 
-use Throwable;
 use Illuminate\Container\Container;
+use Spiral\RoadRunner\Http\PSR7Worker;
 use Spiral\RoadRunnerLaravel\Dumper\Dumper;
 
 if (!\function_exists('\\rr\\dump')) {
@@ -17,7 +17,7 @@ if (!\function_exists('\\rr\\dump')) {
      *
      * @return void
      *
-     * @throws Throwable
+     * @throws \Throwable
      */
     function dump($var, ...$vars): void
     {
@@ -36,7 +36,7 @@ if (!\function_exists('\\rr\\dd')) {
      *
      * @return void
      *
-     * @throws Throwable
+     * @throws \Throwable
      */
     function dd(...$vars): void
     {
@@ -44,5 +44,19 @@ if (!\function_exists('\\rr\\dd')) {
         $dumper = Container::getInstance()->make(Dumper::class);
 
         $dumper->dd(...$vars);
+    }
+}
+
+if (!\function_exists('\\rr\\worker')) {
+    /**
+     * Get the RoadRunner PSR worker.
+     *
+     * @return PSR7Worker
+     *
+     * @throws \Illuminate\Contracts\Container\BindingResolutionException If called outside of RR worker context
+     */
+    function worker(): PSR7Worker
+    {
+        return Container::getInstance()->make(PSR7Worker::class);
     }
 }